% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ryxfcn.R
\name{plot.ryx}
\alias{plot.ryx}
\title{Plotting Correlations between selected variable and other variables in selected dataset}
\usage{
\method{plot}{ryx}(x)
}
\arguments{
\item{x}{An object of class \code{ryx}, generated by the \code{ryx} function.
It contains the correlation, p-values, and level of significance between the selected variable and all the other variables
in the selected dataset.}
}
\value{
A plot object specific to the method. The function returns a visualization of the correlations between
the selected variable and all the other variables in the selected dataset.
}
\description{
Creates visualizations for the correlations between variables
This function works with objects created using the \code{ryx} function.
}
\examples{
# Example usage:
library(MASS)
object <- ryx(Boston, y="medv")
plot(object)
data(mtcars)
object2 <- ryx(mtcars, y="mpg")
plot(object2)
}
