% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ryxfcn.R
\name{print.ryx}
\alias{print.ryx}
\title{Printing correlations between selected variable and other variables in selected dataset}
\usage{
\method{print}{ryx}(x)
}
\arguments{
\item{x}{An object of class \code{ryx}, generated by the \code{ryx} function.
It contains the correlation, p-values, and level of significance between the selected variable and specified
variables in the selected dataset. If none are specified, the selected variable will be
compared to all other numeric variables in the dataset, excluding itself.}
}
\value{
The function returns a formatted print of the correlations between
the selected variable and all the other variables in the selected dataset.
}
\description{
Prints correlations between selected variable and other variables in selected dataset
}
\examples{
# Example usage:
library(MASS)
object <- ryx(Boston, y="medv")
print(object)
# Example that compares selected variable with specified variables from a given dataset
data(mtcars)
object2 <- ryx(mtcars, y = "mpg", x = c("qsec","gear","carb"))
print(object2)
}
