% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ryx fcn.R
\name{ryx}
\alias{ryx}
\title{Calculate Correlations Between Selected Variable and Other Variables}
\usage{
ryx(data, y, x)
}
\arguments{
\item{data}{A data frame containing the dataset to be analyzed.}

\item{y}{A character string specifying the variable (must be numeric) for correlation analysis.}

\item{x}{An optional character string specifying the variable that the user wants to compare variable y to.
If not provided, all numeric variables in the dataset except \code{y} will be used.}
}
\value{
The function returns an object of class \code{ryx}, which includes \code{y},\code{x} ,
\code{df} which is a data frame with the variables of interest, \code{variable} is the column name of the variable,
\code{r} is the correlation coefficient, \code{p} is the p-value for the correlation test
and \code{sigif} is the significance level indicators.
}
\description{
Calculates the correlations between a selected numeric variable
and other numeric variables in a dataset. Users can specify which numeric
variables to include in the comparison. If no numeric variable is specified
the function will compare the selected variable with all of the other variables
in the data set. The function also evaluates the statistical significance of each correlation.
}
\examples{
# Example usage with the MASS package
library(MASS)
object <- ryx(Boston, y = "medv")

# Example usage with the mtcars dataset
library(datasets)
object2 <- ryx(mtcars, y = "mpg")
}
